ci: remove pin of docker images
These pins were added a while ago after seeing a number of pipeline
breakages from docker image updates. We haven't seen such a breakage
in some time but what we have seen is docker image update PRs sit for
a considerable amount of time.

Let's remove the pin. If we see instability in the future, we can
address it then.

Signed-off-by: Steven Danna <steve@chef.io>